remove action controller, some code into action_helper, other refactoring
