+++ acceptance_tests/features/address.feature
@@ -91,7 +91,7 @@ Feature: Address updates
 
   @hardcoded_census_values_for_collection_and_action_plan_ids
   Scenario: Skeleton cases are excluded from action rules
-    Given sample file "sample_1_english_SPG_unit.csv" is loaded successfully
+    Given sample file "sample_1_english_SPG_unit.csv" is loaded successfully with census action plan collection ids
     And a NEW_ADDRESS_REPORTED event is sent from "FIELD" with sourceCaseId
     And a case created event is emitted
     When set action rule of type "P_RD_2RL1_1"
