+++ acceptance_tests/features/steps/ccs_property_listed.py
